+++ NEWS.md
@@ -1,5 +1,8 @@
 ## formatters 0.5.2.9001
 
+### Miscellaneous
+ * Decimal alignment now throws informative error if scientific notation is used.
+
 ## formatters 0.5.2
  * `paginate_to_mpfs` can handle single column tables and listings.
  * Added decimal alignment support with `decimal`, `dec_right`, and `dec_left`.